sistemaGestionUsuario/requirements.txt
djangorestframework~=3.16.0
Django~=5.2.4
django-cors-headers~=4.7.0